+++ jsonrpc/metrics/metrics.go
@@ -55,22 +55,7 @@ func Register() {
 				Help: "[JSONRPC] number of requests handled",
 			},
 			Labels: []string{requestHandledTypeLabelName},
-		},
-		{
-			CounterOpts: prometheus.CounterOpts{
-				Name: requestMethodName,
-				Help: "[JSONRPC] number of requests handled by method",
-			},
-			Labels: []string{requestMethodLabelName},
-		},
-		{
-			CounterOpts: prometheus.CounterOpts{
-				Name: requestWsMethodName,
-				Help: "[JSONRPC] number of ws requests handled by method",
-			},
-			Labels: []string{requestMethodLabelName},
-		},
-	}
+		}}
 
 	start := 0.1
 	width := 0.1
@@ -84,6 +69,7 @@ func Register() {
 	}
 
 	metrics.RegisterCounterVecs(counterVecs...)
+	metrics.RegisterCounterVecs(counterVecsX1...)
 	metrics.RegisterHistograms(histograms...)
 	metrics.RegisterHistogramVecs(histogramVecs...)
 }
